Migrate `*Permission` resources to `infer` (#811)
Migrate the next 2 resources to `infer`.
`pulumiservice:index:TeamEnvironmentPermission` is a drop-in
replacement, but `pulumiservice:index:TeamStackPermission` is not.



Here, `permission` on `pulumiservice:index:TeamStackPermission` stops
being plain. `infer` does not support plain inputs on custom resources,
since there is generally no benefit in doing so. We should not have used
`plain` here. That means this is a minor break, but in the right
direction. Up to you if you think this is a good idea or not.

+++ sdk/python/pulumi_pulumiservice/_enums.py
@@ -192,7 +192,7 @@ class TargetActionType(_builtins.str, Enum):
 
 
 @pulumi.type_token("pulumiservice:index:TeamStackPermissionScope")
-class TeamStackPermissionScope(_builtins.float, Enum):
+class TeamStackPermissionScope(_builtins.int, Enum):
     READ = 101
     """
     Grants read permissions to stack.
